Added output_data when updating Scene.initialization.mdl for dyn geom.
Recent versions of CellBlender now generate MDL in the "output_data"
subdirectory. When the dynamic geometry is exported, it must update both
the main.mdl file and the initialization.mdl file. The output_data
portion of the path had not been included in the second file. This
commit adds it (although the path manipulation is far from clear).

Note that this version doesn't seem to display the dynamic geometry from
the MDL as it did in earlier versions (running with Blender 2.77 for
example). It's not clear where this capability was dropped. It does,
however, appear to generate the proper MDL and dynamic geometry files.
The resulting playback in CellBlender shows that the molecules are
affected by the dynamic geometry even though the geometry itself isn't
being drawn.