Check dynamic objects for being scripted or not.
The "dynamic" property indicates whether an object changes (via either
shape keys or scripting). The "script_name" property is empty when shape
keys are used and it contains an internal text file name when scripting
is used. There is also a global flag indicating whether there is ANY
dynamic geometry in the entire model (for efficiency). This change adds
checking of the script name to determine whether updating the scene is
handled automatically by the shape keys or requires loading new dynamic
geometry from an MDL file.