Change the dynamic geometry exporting to include an intermediate file.
The previous version generated a dyn_geom_list.txt file with an entry
for each object containing the time and the MDL file with geometry:

  0 ./dynamic_geometry/Cube_frame_0.mdl
  0 ./dynamic_geometry/ScriptCube_frame_0.mdl
  1e-06 ./dynamic_geometry/Cube_frame_1.mdl
  1e-06 ./dynamic_geometry/ScriptCube_frame_1.mdl
  2e-06 ./dynamic_geometry/Cube_frame_2.mdl
  2e-06 ./dynamic_geometry/ScriptCube_frame_2.mdl
  3e-06 ./dynamic_geometry/Cube_frame_3.mdl
  3e-06 ./dynamic_geometry/ScriptCube_frame_3.mdl

That version worked fine when there was only one dynamic object. But
MCell does not allow more than one file to be specified at a given time.

This version creates an additional set of files named "frame_###.mdl"
(one for each frame). These additional files include the individual MDL
files for each dynamic object. All of the "frame_###.mdl" files are then
listed in the "dyn_geom_list.txt" file with the appropriate time.

Note that there are still a few unresolved issues with dynamic geometry:

  Partitions are not included in this version.
  Release sites using dynamic objects don't work in this version.
  Scripted Objects don't make use of MDL's Transformation Commands.