Generate and use the TRANSLATE MDL keyword to preserve object origins
The previous dynamic geometry scripting interface assumed that all
objects had their origins at the Scene's origin. This commit adds code
so that scripts may define a different origin for their objects. This
origin is passed to MCell with the "TRANSLATE" keyword in the generated
MDL. The "TRANSLATE" keyword is also recognized when reading back the
code-generated MDL for displaying dynamic geometry in Blender.