Properly handle last frame with scripted geometry
CellBlender traditionally shows iterations from 0 to num_iterations.
The previous code only showed 0 to num_iterations-1 (which is actually
num_iterations frames). This change generates the dynamic MDL for the
num_iterations frame itself and displays it. This commit also restricts
the current from to be less than "scene.frame_end" so the object will
still show if the user scrolls past the end of the iterations in the
time line. This makes dynamic geometry consistent with our molecule
visualization which also show the last frame even when the time cursor
has scrolled beyond the end of the simulation.