Added a dynamic display source control
The dynamic geometry that Blender displays can really come from three
different sources:

  1. Blender's built-in mechanisms (like shape keys)
  2. Reading data files via a frame change callback
  3. Execution of a script via a frame change callback

Previous code could handle the first two cases, but not the third. Even
when a dynamic geometry script was being used, that script first needed
to be called to generate the files (regularized MDL) before the dynamic
geometry could be displayed in Blender.

Note that there may be some performance tradeoffs between reading a
regularized MDL file and generating the geometry via a Python script.
That performance tradeoff could easily change based on the complexity
of the script and the amount of data to be read from file. It's easy
to imagine a case requiring lots of computation where the pre-computed
files would be more efficient. It's also easy to imagine very simple
geometry where the Python to generate the geometry is simpler than the
Python to read it (not to mention the file read time itself). So either
method may outperform the other under certain conditions.

But most importantly, when a script is being developed, it is very
helpful to see the resulting geometry immediately without having to
export the project to dynamic geometry (MDL) files. That reason alone
justifies a mode where dynamic geometry is displayed by calling the
generating script directly rather than forcing it to first generate
files to be read back later.

This commit creates the preliminary set of controls needed to support a
choice between these 3 different dynamic geometry sources.