Added a scripted dynamic geometry example and related changes.
This version includes a new example in scripted_dyn_geo.py. That file is
very similar to the other examples except that it doesn't contain any
line breaks (just one looong dictionary). It should be reformatted to be
consistent with the other examples.

The scripted dynamic geometry example builds a cube at the origin with a
script to change the z-coordinates of the top surface with a sinusoidal
variation. Future improvements might include a concentration clamp and
counting to show sinusoidal molecule count consistent with the
sinusoidal volume changes.