Wrap text for long warnings/errors.
When an MCell simulation finishes running, the stderr is printed beneath
CellBlender's "MCell Processes" list. These warning can be pretty long
and are usually cut off. Often, you need to expand the tool shelf
considerably in order to read the message. To address this problem, we
wrap the text based off of the tool shelf width.

There are other places where we might want to wrap long text, so I'll
try to generalize this process in the next commit.

The code in this commit was based on a stack exchange post found here:

http://blender.stackexchange.com/questions/74052/wrap-text-within-a-panel