This version can export the data model found in a .blend file.
A CellBlender .blend file contains an ID string property pickle of the
data model when it was saved. This change adds operators to save that
data model to a file when a .blend file is opened that doesn't match the
current CellBlender ID. This is useful when "downgrading" a .blend file
that was saved with a newer version of CellBlender. This "downgrading"
must be done manually by changing data fields and the data_model_version
fields to match an earlier version of CellBlender. Note that this commit
doesn't add the geometry to the internal data model that it exports. So
data models exported with this commit will not have any geometrical
objects.