Added the geometry and materials by constructing a full data model.
As mentioned in the previous commit, the data model stored in the .blend
file does not contain geometry (or materials) since these are already
stored in Blender's data. But when a .blend file is being downgraded to
work with an earlier version of CellBlender, it's important to export
that geometry (and materials). This commit adds those features to the
export functions when a version mismatch is detected.

Along with the previous commit, this provides a mechanism to manually
downgrade a data model. The procedure is:

1. Open the .blend file in CellBlender (any version after this commmit)
2. When the mismatch is detected, use one of the export buttons to save.
3. Use a data model editor to change fields to match an earlier version.

In many cases, changing the data_model_version to match the earlier
version should be sufficient (when fields had been added that were not
present in the earlier version). In that case the extra fields added for
the newer version will be ignored by the older version.