Added notes about deleting Shape Keys to Shape Key example.
The Shape Key example (without a .blend file) works well, but suffers
from an inability to remove existing shape keys. So while it can be run
multiple times, it can't be re-initialized (from the Examples button)
without first resetting to defaults (or otherwise deleting the existing
shape keys).

This commit doesn't solve the problem but introduces some comments and
references that might be helpful in solving it. There have been a number
of posts on BlenderArtists discussing the problem that are referenced in
this commit.

Various attempts were made to use bpy.ops.object.shape_key_remove, but
they resulted in context errors. So the code is currently commented out.