mproved viewing of Shape Key example and tried to remove shape keys.
Made a few small changes to get the view to show the fully stretched
cube in the display when initialized. Continued to try to remove the
shape keys if they already exist, but it fails to do so.

Wrapped the "shape_key_remove" function in try/catch so that it makes
an attempt (even though it fails).