Temporary fix to dynamic geometry shape key example.
The dynamic geometry shape key example creates shape keys using:

  bpy.ops.object.shape_key_add(from_mix=False)

Those shape keys are attached to the object, but they're also stored in
the global bpy.data.shape_keys location. When the example is loaded, it
should remove all the shape keys from bpy.data.shape_keys, but existing
code was unable to do so (see multiple commented sections in the file
cellblender_examples.py).

The temporary fix is to simply use the most recently created shape key
and ignore the ones that accumulate in bpy.data.shape_keys. If the
example is loaded 6 times consecutively, then they will be:

  ['Key', 'Key.001', 'Key.002', 'Key.003', 'Key.004', 'Key.005']

The temporary fix will use the last one ('Key.005' in this example).
Note that this accumulation only happens each time the example is loaded
from the Examples panel and not each time it is run as a simulation.

The multiple approaches in the commented code contain references to
possible solutions, but none were effective. Those comments should be
removed when a solution is found.