Improved initial view of the scripted dynamic geometry example
The scripted version does not currently draw dynamic geometry while it
is being loaded (in cellblender_examples.py). So the timeline cannot be
used to draw the full-sized version for centering the view (as is done
with the shape key version).

Instead, a trick is used to create a temporary second object at the top
of the full-length cube. The view is then centered (view_all) with this
temporary object in the scene. The temporary object is then deleted
leaving the view centered on what will be the full range of the dynamic
object.

Also, the molecule count was reduced to speed up the simulation a bit.
This was done at both the concentration clamp and the initial release
(so the initial count roughly matches the requested concentration). This
version is still slower than the shape key version because the shape key
version has more molecules, doesn't use a concentration clamp, and
doesn't plot any results. It's not clear whether it would be best to
make these two versions identical or to have them show different
features (like the concentration clamp in this example).