Disabled Python version that uses libMCell ... doesn't really work.
The sim_engines_and_runners branch doesn't need to implement all
combinations of engines and runners at this time. The Python version
that uses the C++ libMCell has had some SWIG-related problems and isn't
needed to demonstrate the separate engines/runners concept. So it has
been disabled in this version.