Minimal MCell3 engine implemented and working.
This version can run MCell, Smoldyn (limited functionality) and both the
pure Python and pure C++ simulators (also limited functionality). They
all produce molecule visualization data for CellBlender.

This version relies heavily on run_data_model_mcell_3.py which was
borrowed from mdl/run_data_model_mcell.py.

Right now, the "prepare_runs" function also runs the model. This needs
to be factored out to produce the list of commands to pass to the
runner.

Also, with thoughts of running on remote / distributed machines, it
might be best for the engine to pass file lists to the runner along with
the commands. The file lists might be input or both input and output.
This would be in addition to the command list to run.