Updated cellblender_reaction_output to handle the new data layout.
Previous commits updated the visualization code for the new layout:

  /DIR          (used for an abritrary directory)
  /FILE_TYPE    (either "react_data" or "viz_data")
  /SEED         (a seed number directory)

This commit updates the reaction output code for that same structure.

Note that the newer data layout files (data_layout.json) have a version
number. If the version is not found or not "2" then the file is
processed with the old format. If "2" is found, then the definitions
above are used.