Simple autodetection of Linux platform in CellBlender's main makefile
Our makefiles make a lot of assumptions that we're running on a *nix
platfrom. This change doesn't fix that, but it does allow for a
different install directory based on whether the "uname" command returns
"Linux" or not.

More work could be done to further investigate the best way to do this,
and there's a good reference at:

  http://stackoverflow.com/questions/714100/os-detecting-makefile

But for now, this should save the MacOS users from constantly editing
the CellBlender make file.