Added bounds checking for all objects via a check box.
Smoldyn requires a bounds for its simulation. The bounds can be set via
the Smoldyn pluggable GUI to any range desired. However, in practice,
the bounds are implied by the objects in the simulation. This flag gives
the option of setting the bounds automatically based on the maximum
bounds of all objects.