Smoldyn Engine: Calculate bounds of each object for internal point.
Smoldyn requires an internal point for each object to define a
"compartment". MCell uses the face normals to determine inside and
outside, so no such "internal point" is needed.

Since the calculation of bounds was also needed, that code was adapted
to find the bounds of each object first. For convex objects, this should
provide an easy way to calculate an internal point.

Note that according to the Smoldyn documentation, "the region of a
compartment is everywhere in space from which one can 'see' the
interior-defining point using a straight line, without crossing any of
the bounding surfaces". This may require multiple internal defining
points if the object is non-convex in many cases anyway. Satisfying that
requirement for non-convex objects will require more work anyway, so
this is a good first step. A better algorithm might define points based
on negative face normals.