Updated some of the random motion calculations in engines to match MCell
Some of the  gaussian random number generators in some of the engines
were intentionally temporary just to get something running. This commit
corrects them to be consistent with MCell. Additionally, limited C++
version also accepts a "test_rng" parameter which will print the mean
and variance of 1000 gaussian samples (should be close to 0 and 1).

Note that they still do not generate the same random sequences as MCell.
They are not indended to produce bit-for-bit identical output.