Updated the dynamic engine/runner code to remember settings.
The dynamic nature of the engine and runner code currently requires some
execution at startup with a valid context. The load_post handler's
context seemed to be "None" so it wasn't usable. As a workaround, this
version triggers the reloading of the dynamic code when run. This
version also retains some commented load_post code in case it's useful
in the future. It will be removed in a subsequent commit, but it's still
available in this one (although commented out).