Removed some exploratory code and only check for changes when None.
As mentioned in the previous commit, some of the exploratory code was
removed in this commit to minimize the amount of commented code.

This version checks whether the active engine and runner are None before
calling plugs_changed_callback. This handles the case of running after
restarting Blender. It will reload the appropriate modules before
running them, but the panels will appear empty until the run (or
refresh) is pressed. Taking either action will repopulate the panels as
they were before being saved.

This commit makes it much easier to run a dynamic engine or runner after
restarting Blender, but it doesn't help when a version change reloads
from the data model. That's because none of the dynamic engine/runner
properties are stored in the data model at this time. The code has been
considered too experimental to justify modifying the data model yet.