Added code to support surface classes and surface regions.
This commit generates MDL for surface classes and surface regions
defined in CellBlender.

There's a rumor that MCellR may require some kind of surface class on
all surfaces. If that's true, then the exporter should add a default
reflective surface class. For this commit, those surfaces would have to
be added manually in CellBlender since it doesn't autogenerate anything.