Corrected shape exporting when using solid geometry subtraction operator
The data_model_to_mdl_3.py had been exporting the "shape" value for
release objects by prepending a "Scene." string. This works fine as long
as the release object is a single object and not the result of a solid
geometry operator (like ObjA - ObjB).

This is a quick fix to prepend the "Scene." string to each object in a
solid geometry expression using the subtraction operator (-). It will
take a little more work to handle other operators as well.

This problem was found when working on the mcell3r engine in the bngl
branch. It's being applied here to the source of that code in the mcell3
engine.