Changed the membrane definition to just plain membrane name (breaks!!)
The first (previous) versions used a membrane definition that was
essentially raw MDL for the membrane name field. This was a temporary
measure until the proper specification could be found.

So a previous definition for object CP of:  PM OBJECT CP[ALL]
                       now becomes simply:  PM

This will break previous models, but since this feature has only been
supported for a few days, it shouldn't be too painful.