Re-enabled reaction checking when not in BioNetGen Mode.
Note that reaction checking retains its state. This means that if a file
with BNG reactions is opened when NOT in BioNetGen Mode, the reactions
will be flagged as errors. Setting BioNetGen Mode will NOT the clear the
errors unless a field is updated in each reaction (this is typically
accomplished by clicking in one field in the reaction and pressing enter
without changing the value).

It's not clear how this should be handled. For now, the BioNetGen Mode
will likely be stored as a preference. But it might be better stored as
a property of the model and saved in the Data Model. This may also
depend on our future ability to detect BioNetGen data in various fields.