Re-enabled release site checking when not in BioNetGen mode.
As with the other recent changes, this commit checks the state of the
cellblender_preferences.bionetgen_mode before doing checking. In most
cases, the BioNetGen checking is just a "pass" to be implemented later.

The bngl branch has been merged with the development branch. It should
have additional functionality when the bionetgen_mode is enabled, but
there is no checking of anything that requires understanding of BNGL
(such as molecules, reactions, release sites, etc).

This version passed the test suite except for the following tests:

  Java Runner   - Can't be run because Java is now a dynamic runner.
  OpenGL Runner - Can't be run because OpenGL is a dynamic runner.
  Torus Tests   - It is suspected that the vertex numbering differs.

The Java and OpenGL Runner tests should be re-coded to test those
runners via the new dynamic interface (they don't exist under the old
static interface that the test uses.

The Torus tests could benefit from our own code to write build stable
geometry that's known to be the same between Blender versions.

This is a release candidate for the 2017 Workshop. It should support
our "normal" molecules when bionetgen_mode is disabled, and it should
support BNGL syntax molecules when bionetgen_mode is enabled.