Added automatic postprcessing to MCell3R engine.
Technically, any post-processing should be passed on as "jobs" for the
runner to do (or be part of the individual jobs passed to the runner).
But since the MCell3R engine actually does the running, it's more
convenient to have it run the post-processing automatically within the
engine for now.

This commit makes that change.