Began separating engine and runner functionality in the MCell3 Engine.
The current MCell3 Engine exports the model and runs it via either the
MPP library or the SGE code embedded in that engine. Rather than passing
an actual command string, the parameters are passed as a list which must
be interpreted by the runner as parts of an MCell command line to be
assembled into the actual command line.

In order to separate the functionality, the runner should just get a
list of command strings rather than a list of command components to be
assembled. This first step just moves the ".main.mdl" portion of the MDL
filename option from the "runner" into the "engine". The runner
shouldn't need to know to add ".main.mdl" as part of its command string.