Added support for a custom "draw_layout" function for engines/runners.
The Queue runner (and possibly others) may want to do custom panel
drawing in addition to (or instead of) the default parameter_dictionary
drawing. This commit calls any "draw_layout" function found in a runner
(or engine). In order to be consistent with other "draw_layout"
functions, this version uses the common signature:

   def draw_layout ( self, context, layout ):

The "self" will be "None" since it isn't a subclass of bpy.types.Panel.
The "context" and "layout" will be as expected.

Note that the "self" may be removed in future versions or possibly
replaced with an actual Panel if that's useful.

Using this approach, the current Queue runner might be able to be coded
as a pluggable runner with all of its current functionality. But there
are a few other potential problems (such as operator support) that are
not resolved yet.