More code to support a dynamic Queue Runner.
This version checks to see if there's a "get_pid" function in any
dynamic runners to enable the PID portion of the display. This may not
be the best way to check, but it's a start. Note that the PID portion
still isn't being drawn with the dynamic Queue Runner yet. Needs more
work.

This commit shouldn't affect any non-dynamic functionality, so while it
doesn't work yet, it's still usable when using non-dynamic runners.