Moved the dynamic engine/runner panels to top when shown.
The compact drawing takes up less room and makes the engine/runner
choices more evident. The current version draws them only when the
"Dynamic" option is chosen, but eventually all engines and runners
will be dynamic, and the "Dynamic" option will be gone. The list of
runs will appear between the "Export/Run" button and the Output and
Control panel as it does with the Queue Runner.