Renamed mcell3 engined directory to mcell3dm (for MCell 3 Data Model).
The existing MCell3 engine uses a Data Model interface rather than
direct exporting of MDL from CellBlender's internal properties. With the
new engine/runner system, both kinds of exporters should be supported.
In other words, we should be able to move the current (internal) MCell
exporter into an engine module that exports MDL from Blender's
properties rather than from a Data Model. We may also want an engine
that operates directly from the Data Model as the current mcell3 does.

With those thoughts in mind, I've renamed the current mcell3 engine to
be mcell3dm to reflect that it operates off of a data model. This will
allow creation of a new "mcell3" engine that exports directly from
Blender's properties. This will essentially be the current MCell3
exporting code repackaged as a simulation engine.

The general philosophy is that multiple engines can be supported, and
they can get their data from potentially different sources. There are
advantages to writing an engine to work with a data model because that
engine could then operate outside of Blender (in Galaxy, for example).
Conversely, there are also advantages (such as performance) to writing
an engine that uses Blender's properties directly. Since we already
have an existing property-based MCell exporter, it makes sense to use
that as an example of such an engine.

Finally, there's no prohibition to an engine that supports multiple
data sources (such as Data Model AND Blender Properties). Such a runner
might display a flag for the user to choose a data source. At this time,
however, it's easier to keep them separate to simplify the coding.