Updated the Java, OpenGL, and CmdLine runners for string or dictionary.
There may be cases where it's easier to pass a full command string
rather than a dictionary of command parts for each job to run. This
commit updates the Java runner, OpenGL runner, and Command Line runner
to accept either strings or dictionaries in its command list.

The Engine/Runner interface is now defined by a list which contains
one string or dictionary per job as follows:

  If a list item is a string, it is assumed to be a complete command.
  If a list item is a dictionary, it will contain the following keys:
    'cmd'    - Mandatory entry with the command to run (args optional)
    'args'   - Optional arguments to be passed along with the command
    'wd'     - Optional working directory where command is executed
    'stdout' - Optional stdout: None=/dev/null, ""=Console, "name"=File
    'stderr' - Optional stderr: None=/dev/null, ""=Console, "name"=File

Note that not all features are currently implemented by all engines or
runners.