The Dynamic Queue runner is mostly working in this version.
The Dynamic Queue runner currently has 2 buttons to register and
unregister the operators that it shares with the non-dynamic version.
These are a temporary workaround and they are used to unregister the
existing (typically static) Queue runner operators, and then to register
the dynamic Queue runner operators. If this isn't done, then the static
Queue runner operators will be called since they're loaded by default.
This should be handled automatically in the future.

Right now, the run progress is handled by checking stdout for the
signatures of the different engines, and handling ones that match.
This would be better handled via a callback in each engine that takes
the stdout string as a parameter and returns the progress that it finds
since it should know how to read the output for its own engine.

Note that Smoldyn shouldn't be run with the Command Line runner because
it sometimes prompts: "Overwrite existing output file ... (y/n)?" which
can't easily be answered. This hangs the Command Line runner.

Also note that Smoldyn appears to generate a message on stderr when
it completes normally. This shows up as an error in the Queue Runner
panel, but doesn't cause any other problems so far. This might also be
handled by a callback to the engine to process stderr and return an
appropriate message for CellBlender to display.