Working on better Queue runner support for heterogeneous engines.
The previous interpretation of the process_list entries assumed a
standard QueueRunner/MCell format. But with the new dynamic engines and
runners, this assumption is no longer valid.

This commit introduced an optional "get_progress_message_and_status"
function in simulation engines. If such a function is present, it will
be called with the stdout_txt that had been previously processed in the
queue runner itself. This allows each engine to write code to process
its own simulation output. So a simulator that prints when it releases
molecules could have those releases displayed by its own engine code
that knows how to find and display those events in the stdout buffer.

This version also introduced code to unregister and re-register the
engines and runners when changed. This will be performed for those
engines and runners that contain a "register_blender_classes" function
and an "unregister_blender_classes" function. This allows dynamic
engines and runners to declare Blender Operators that need to be
registered. At this time, however, the re-registration only takes place
on a change of engine/runner and not when Blender starts up.