Added get_progress_message_and_status to engines and removed hard-coded
The previous version had hard-coded status handling of several known engines.
In other words, it looked for certain signatures in their stdout history
and used it to decode the progress information.

This version removes that hard-coded searching from the Queue Runner and
instead calls the "get_progress_message_and_status" function in the
engine (if one exists). The result is displayed after the PID in the
processes list. Any engines that don't provide a progress function will
be displayed with their PID followed by their name as shown in the
engine selector.

This version implemented a custom get_progress_message_and_status
function for:

  Limited C++
  MCell3 Data Model
  MCell3

The Limited Python engine already supported this function in the
previous commit (where it was prototyped).

Note that the Limited C++ simulation doesn't include the total
iterations in its line-by-line output, so the progress function returns
a string with raw iterations rather than a percentage. This shows the
flexibility of this interface when working with existing simulators. It
also suggests the possibility of other status that might be available
from the simulations. It might, for example show the total number of
molecules, or the current iterations per second, or show recent events
like molecule releases or reaction summaries. In other words, it's up to
the engine writer to provide whatever status might be useful and is
available from the simulator itself.