Added a Sun Grid Engine runner that works with most of the engines.
Added a runner module that supports submitting jobs to a Sun Grid Engine
host. This runner is a modified version of the current MCell-only SGE
runner, but it supports running other simulations. It's been tested with
MCell3DM, C++ Prototype, and Python Prototype. It hasn't been tested
with Smoldyn yet because the Smoldyn engine will likely need to disable
graphics output (an option not yet in the Smoldyn engine).