Added a "developer_mode" preference for options that are under development.
Since we don't release CellBlender very often, we often encourage users
to use the development branch to get the latest features. However, some
new features in the development branch may be broken at any point in
time. The "developer_mode" preference is a boolean flag that is intended
to enable these "in progress" features so they can be developed and tested
in the "development" branch without exposing them to non-developers who
may be using that branch for non-development purposes.

In general, any new feature that isn't working yet should only be shown
when the developer_mode is True.