Began work on a fully dynamic version of the run panel.
The "Show Developer Options in Panel" flag is being used to swap out
the drawing of the old run panel and the new run panel. This allows
prototyping the new panel without having to work around the old panel.

Also added support for saving text logs and deleting text logs in the
new dynamic Queue runner (although it's a temporary solution).

The dynamic Queue runner still uses a number of functions from the
static Queue runner which should be duplicated in the dynamic version
before the static version is removed.