Arranged the main panel for consistency and to save a row.
The Examples and PBC panels now share the top row to save a row and keep
the rest of the panel layout consistent with previous versions. We might
want to review the panel layout from time to time by the group.

Also increased the "show_button_num" BoolVectorProperty for the
additional PBS button.