Cleanup of Preferences area, Addition of Handy Blender Settings
The Preferences area had a mix of CellBlender and Blender properties.
This commit added a new subsection (with Blender icon) for purely
Blender settings. Blender settings now include background color and
wire color - which are handy for quick viewport renderings. Note that
because some of these (background color and wire color) are Blender
"Preferences" and not Blender "Properties" they are not saved in the
.blend file and apply to all files that might be opened. These must
be saved as preferences to keep them across Blender sessions. It might
be handy to have a "Save Blender Preferences" button as well (maybe in
a subsequent commit).