Set data_model[' mol_viz']['file_dir'] to a blend file relative path.
Our policy has been that all paths to generated data files should be
relative to the .blend file. However, the mol_viz['file_dir'] had been
saved as relative to the working directory. This was generally not a
problem when running from the same working directory, but might cause
problems with a moved .blend file. Note that this commit fixes the path
stored in the data model but doesn't change the way it's represented in
Blender's properties. So the display, for example, will still show a
long path rather than the short relative path in the data model.