Got MCell 3 working as an engine with dynamic geometry.
This version was tested with all non-MCellR engines and worked.

This version was able to run MCell3 with dynamic geometry.

Note that much of the MCell3 code is called via operators from the same
locations as before. This version hasn't moved the MCell3 export code
into the engine location yet.