Dynamic export of MDL from properties seems to work.
This version "appears" to use the dynamic MCell3 exporter to export and
run a model. However, some of the functions in the model are used by
CellBlender's main __init__.py file for import and export menus. For
that reason, the io_mesh_mcell_mdl directory was not yet removed. It's
not clear at this time how those menu options should be supported. It
might be part of the engine API to provide a number of menu functions.
But that's still not determined.