Moved much of the MCell panel drawing into the MCell3 Engine
The panel now draws itself from a combination of parameter_dictionary
and draw_layout(). The parameter dictionary is drawn first, and then any
custom Blender drawing is done via the "draw_layout" function. In this
case, the "draw_layout" gets most of its data from CellBlenders
properties (as it has in the past).

Note that the parameter_dictionary still isn't stored as a persistent
Blender property, so it needs to be reinitialized every time CellBlender
is restarted.

Things to do:

  Shared job list panel (common format for all runners?)
  Persistent state of Engine and Runner parameter_dictionary
  Moving data between hosts (rsync or similar)
  Removing old static code (io_mesh_mcell_mdl and other runners)
  Policy on finding the proper MCell (engine panel or global?)