Implemented a simple rsync file transfer to and from remote host for SGE
This version updates the SGE runner to provide a remote path and a
remote user. If the remote path is empty, it is assumed that the file
system is shared between the local machine and the host. In that case,
no file transfers will be performed. But if the remote path is not
empty, then the blend_file_name_files directory will be transferred to
the remote system via rsync prior to the run. Bringing files back
currently requires pressing the "Fetch" button in the runner panel
because this version does not detect completion of the runs on the
remote host.