Added the ascii_enable flag in new io_mesh update to the MCell3 engine.
Commit a2ceb576fcafe07... added an "ascii_enable" flag to the VIZ panel
and checked it in the io_mesh_mcell_mdl code when exporting. This flag
somewhat duplicates (in a different way) the existing "export_all_ascii"
flag in the Advanced Options panel.

This commit changes the mcell3 engine to use the new "ascii_enable" flag
in the same way that it is now being used in the io_mesh_mcell_mdl code.

It's not clear which functionality (or both) should be used or where the
flag belongs in the panels.

Note that the new flag doesn't appear to be saved in the data model. The
older "export_all_ascii" is in the data model's Initialization section.