Began adding host selection controls to SGE runner
It is anticipated that some remote systems may not have a shared file
system. So the rsync that moves files might have to know which hosts
will need to get the files. That may require a dynamic hosts list as
used in the hard-coded SGE runner. This is best done with real Blender
properties rather than the simplified dictionary interface provided to
most Engines and Runners.

This commit has moved much of the hard-coded SGE host selection code
(including a Blender template list) into the SGE Runner. While this
commit changes the SGE runner, it might have been better done to a new
runner to preserve the older simpler interface. That will be done in a
future commit.