Discontinued work on Host-Selectable SGE Runner, plug_active = False
As we develop different engines and runners some may not be ready for
general use. This change adds a check for "plug_active" in each module
to be used to either show or not show that engine/runner in the user
menu. Note that this currently requires changing the source code to
enable or disable individual engines and runners. It might be nice to
have a global flag that can show modules that are works in progress for
developers, but this is sufficient for now.

Also renamed the "Simple" SGE runner back to "Sun Grid Engine" since it
is the best working version. The host-selectable version will show up
with the name "SGE Host Select" when it is enabled.

Finally, the need for a host-selectable version was in anticipation of
clusters that don't have a shared file system. Since this doesn't seem
to be common in our target environments, it is currently on hold.