First cut at putting engines into the data model.
This is a first attempt at storing the engine data into the data model.
It mostly works, but properties are only built from the data model when
the CellBlender ID of the blend file doesn't match the CellBlender ID of
the addon. So while this functionality might be needed eventually, it
still doesn't preserve the engine settings between runs of Blender.

Note also that there was a mysterious error that was produced when
reading back dictionaries that contained function references ... even if
those functions were never called. It appears that Python was trying to
match those function references (raw addresses?) to functions in the
current running version. So rather than ignoring them on reading, those
function references are removed before storing.

Note that this version only handles engines (and not runners).

Also note that since this only works on upgrades, another implementation
will be required for restarting Blender without any version changes:

  Maybe save a pickle of the parameters?
  Maybe look for "save_as_string" and "restore_from_string" in the API?
  Maybe do both?

The pickle or string would be stored as a regular Blender property and
used to repopulate the dynamic properties when a Blend file is opened.
This may be a new technique that we haven't tried before.